requirements.txt
Django==4.0.6
psycopg2-binary==2.9.3
gunicorn==20.1.0
sqlparse==0.4.2
asgiref==3.5.1
python-dotenv==0.20.0
django-filter==22.1
djangorestframework==3.13.1
pillow-9.2.0
drf-base64==2.0